Use create-react-app
sec. 19, video 189

https://github.com/facebook/create-react-app

Installation
$ npm i -g create-react-app@2.1.8

Create a new project
$ create-react-app project_name

It will generate a new directory on the path and populate it with all
of the boilerplate files. Then it's going to install all of the npm
modules necessary. It will also initialize a git repository.

Note the we don't have a Babel config file and webpack config file in
the project. Create-react-app is indeed using both of those tools behind
the scenes. But when we use the boilerplate, all of that config is
abstracted away, allowing us to focus on writing application.

react-script is an npm module specifically designed for create-react-app
apps. This module gives us access to all of thing we can do including
the development server.

Note that the script "eject" is not something we should run because we
can only run it once. When we use eject, we're essentially telling
create-react-app that we want to take over control. That means it's
going to take all of the configuration details that are currently
abstracted and it's going to give you access to them. So we can
customize webpack and Babel. But there should be no need to actually do
this. Though it's available if we feel like we've outgrown what
create-react-app can do.

In the public folder, the manifest.json is used to create what's known
as a progressive web application.

What we're going to do is break this down into a very bare bones
version.

Inside of index.js, we remove the import of index.css and App.js.
In the src folder, we remove index.css, logo.svg, App.css, App,js and
App.test.js.