Introduce the useState hook
sec. 19, video 190

https://reactjs.org/docs/hooks-state.html
https://reactjs.org/docs/hooks-reference.html#usestate

React hooks tie in to our functional components. Stateless functional
components have some limitations. They can't manage local state and tap
into other react features like life cycle methods. In the past, if we
want to access the state in a functional component, we need to refactor
it into class-based component.

A react hook is a function that lets you tap into a react feature like
state or life cycle methods. One of built-in hook, useState, allows us
to use component state in our functional components.
useState accepts a single argument, our initial state value. The type of
the state could be anything, doesn't have to be an object.
It returns a stateful value, and a function to update it.
The setCount function is used to update the state (count in our case).
It accepts a new state value and enqueues a re-render of the component.