Create complex state with useState
sec. 19, video 192

Target: Using useState to build a Note App. We will create a <form> with
an <input> for the title, <textarea> for the body and a <button> to
submit the note.
Then we will show all the notes on the page. We can delete the note with
a <button> under its body.

Remember when we work with arrays of JSX, we need to specify that key
prop.