Introduce the useEffect hook
sec. 19, video 193

Ref: https://reactjs.org/docs/hooks-effect.html

useEffect allows us to do something in functional components. This is
kind of a replacement for lifecycle methods in our class based
components (you can think of useEffect Hook as componentDidMount,
componentDidUpdate, and componentWillUnmount combined.).

In React class components, the render method itself shouldn’t cause side
effects. It would be too early — we typically want to perform our
effects after React has updated the DOM.
This is why in React classes, we put side effects into componentDidMount
and componentDidUpdate.

useEffect is something we call and we pass to a function. The function
is going to run once right away with our default state values
(componentDidMount) and after changes to our component state or props
(componentDidUpdate).
For example, we can use it to synchronize the count state with the
document title.

For <NoteApp>, we can read notes data from localStorage (default to
empty array) and call useEffect to update localStorage when note array
changes.