Add useEffect dependencies
sec. 19, video 194

https://reactjs.org/docs/hooks-effect.html#tip-optimizing-performance-by-skipping-effects

In the last section, the function passed to useEffect runs for every
state update include the input text, which is not what we want. We just
want it runs when 'count' variable changes.
useEffect hook allows us to specify the things we care about by passing
an array as the 2nd argument.

https://reactjs.org/docs/hooks-effect.html#tip-use-multiple-effects-to-separate-concerns

Just like useState, we can use the useEffect hook as many times as we'd
like to in a functional component. Hooks let us split the code based on
what it is doing rather than a lifecycle method name. React will apply
every effect used by the component, in the order they were specified.
To make the function run once when the component first mounts but never
runs again, we can pass an empty array as the 2nd argument. Because it
depends on nothing. It is useful for fetching or reading data. It is a
complete mirror of a componentDidMount.

For Note App, we can add dependencies to existing useEffect call. It
will only run when 'notes' changes. Then use useEffect to load the data
from localStorage (only runs once). This is the sort of pattern we would
use if we were fetching our data from database.