Cleans up effects
sec. 19, video 195

https://reactjs.org/docs/hooks-effect.html#example-using-hooks-1

If our effect returns a function, React will perform the cleanup when
the component unmounts. It allows us to run some code just after the
component is removed.

In this example, we separate the <Note> from <NoteApp>. In <NoteApp>, we
render an array of <Note> by passing props into it.

Note that if we didn't specify the 2nd argument as an empty array for
useEffect in <Note>, the console.log('Setting up effect!') runs even
when the parent component <NoteApp> re-renders.