Introduce the useReducer hook
https://reactjs.org/docs/hooks-reference.html#usereducer

An alternative to useState. Accepts a reducer of type (state, action) =>
newState, and returns the current state paired with a dispatch method.
useReducer is usually preferable to useState when you have complex state
logic that involves multiple sub-values or when the next state depends
on the previous one.
Note that useState actually use useReducer behind the scenes.
We need to define a reducer function before we can call useReducer.
To initialize useReducer state, we can pass the initial state as a 2nd
argument.
We can dispatch an action to run the reducer and manipulate the state.