Introduce the Context API & useContext hook
sec. 19, video 198

https://reactjs.org/docs/context.html?no-cache=1

Context provides a way to pass data through the component tree without
having to pass props down manually at every level.
Context is designed to share data that can be considered 'global' for a
tree of React components, such as the current authenticated user, theme,
or preferred language.

How to use

https://reactjs.org/docs/context.html#reactcreatecontext
https://reactjs.org/docs/context.html#contextprovider
https://reactjs.org/docs/hooks-reference.html#usecontext

1st, we create a new context, NotesContext. This is an object that needs
to be accessible in the component that's provide things and the
component that's consuming things.
That's why we have it in its own file so both of those files can import
it.
createContext doesn't require any arguments. If we provide one, this
would be the default value for the context.

With Context.Provider, we can provide the context's value to anyone in
it and their children and children's children who wants to consume it.
In our case, we share 'notes' & 'dispatch' function.

When we want to consume, we use the useContext hook to access the value,
in our case is an object contains 'notes' & 'dispatch'.

Note that originally we created removeNote function in <NoteApp> and
passed it down 2 level to <Note> as a prop. With Context, we can
dispatch actions in <Note>.
This is similar to what we saw with Redux. Any component in our
application could get access to dispatch and dispatch some action to
change all of the data.