Create custom hooks
sec. 19, video 200

https://reactjs.org/docs/hooks-rules.html
https://reactjs.org/docs/hooks-custom.html

A custom React hook is a function we defined, which uses the built-in
React hooks behind the scenes.

In this example, we need the ability to know where on the screen the
mouse cursor is. There are 3 steps to implement this feature:
1. Setup state to track 'x' and 'y' position (useState)
2. Setup event to listen for mouse movement
3. Remove event listener if unmounted (useEffect) to avoid wasting
   resources

If we implement this feature in our component, we will have 2 limitation
with above approach:
This code is not reusable. If another component in the application also
need the same ability, we need to add this code over again.
The 2nd one is that the component really doesn't need to be concerned
with the implementation details of this feature. All it need is a
reliable updated 'x' and 'y' position.

Implementation

Add a new folder for custom hooks.

Starting a custom hook off with 'use' is a naming convention. Because
'x' and 'y' always change together, we store them as an object instead
of breaking it out into multiple useState calls.

Note that the hooks from React that we're using are not defined in our
functional component but they are defined in a function that our
functional component calls. That's what makes a custom hook.

Be careful not to register the eventListener more than a single time.
That's why we set it up in useEffect with empty dependencies.

Now we can use our custom hook, useMousePosition, in our components.